requirements.txt
wheel
termcolor
schedule
TTS
prettytable
webdriver_manager
selenium
g4f
moviepy<2
imageio-ffmpeg
Pillow==9.5.0
yagmail
assemblyai
srt_equalizer
undetected_chromedriver
platformdirs
huggingface_hub
transformers
